patch : change datatype from int to long in productId

+++ E-Commerce.DAL/Repository/Interfaces/IProductRepository.cs
@@ -10,7 +10,7 @@ namespace E_Commerce.DAL.Repository.Interfaces
 {
     public interface IProductRepository : IGenericRepository<Product>
     {
-        public Task DecreaseProductQuantityAsync(int ProductId, int quantity);
+        public Task DecreaseProductQuantityAsync(long ProductId, int quantity);
 
     }
 }
